% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Concrete.R
\docType{data}
\name{Concrete}
\alias{Concrete}
\title{Concrete Compressive Strength Data Set}
\format{
A data frame with 1030 rows and 8 covariate variables and 1 response variable
}
\source{
https://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength
}
\usage{
Concrete
}
\description{
Concrete strength is very important in civil engineering and is a highly nonlinear function of age and ingredients. This dataset contains 1030 instances and there are 8 features relevant to concrete strength.
}
\examples{
data(Concrete)
}
\keyword{datasets}
