% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Sunspots.R
\docType{data}
\name{Sunspots}
\alias{Sunspots}
\title{Number of sunspots, 1770 to 1869}
\format{
Number of sunspots, 1770 to 1869
}
\source{
Brockwell, P. J., & Davis, R. A. (Eds.). (2002). Introduction to time series and forecasting. New York, NY: Springer New York.
}
\usage{
Sunspots
}
\description{
Number of sunspots, 1770 to 1869
}
\examples{
data(Sunspots)
}
\keyword{datasets}
